Do not add SQL keywords multiple times to auto completion
When performing certain actions like opening a database file the keyword
list for the auto completion of the SQL text editor gets updated. This
was not only done in a very inefficient manner however but also the list
of keywords was not cleared before adding them again. This wastes memory
and makes the editor even more inefficient.