Only show numbered sort indicators when more than one column is sorted
I think we tried this approach before and decided to always show the
numbers to  give a hint to the user that it is possible to sort by
multiple columns. But it does look kind of odd this way because it makes
you look for another sorted column.

See issue #2913.