Fix repositioning of filter widgets after the second click on the header
This happened only for the custom dark/light styles.

The new connects fixes the issue for me. This was discovered by trial and
error, because I don't know what is really causing the problem (seems
some bug relating with event ordering and maybe a race condition).

Second change is irrelevant for the issue, just for optimization of the
loop.

See comment:
https://github.com/sqlitebrowser/sqlitebrowser/issues/2558#issuecomment-1003773910