Edit Database Cell: fixed potential for data loss and better feedback
Remaining detected case for a potential data loss in the cell editor have
been fixed:
- Switching to another main tab.
- Start a filter.

In both cases a dialog for saving or discarding is open.

Additionally, the cell editor shows the cell coordinates currently open.
And the NULL text in margin is reset when the user types in the editor.

See issue #2488