Project files: fix crash when tables does not match DB file
There are project files with settings saved for a table with empty schema
and name.

We let the user load a new DB file when the one referenced by the project
is not found, opening the door to load a DB with a different structure.

A final case for this inconsistencies can be that we let the user cancel
the opening dialog, and the project can be opened without a DB file.

See issue #3007