Optimise executing large SQL files in the Execute SQL tab
This speeds up running very large SQL files via the Execute SQL tab
significantly. Before this commit all the remaining statements for
execution were analysed in order to find out the type of the next
statement. Especially because comments were stripped from the entire
text this would take a long time for large files. With this change the
next statement is prepared for execution first, then the prepared
statement only is analysed, and then it's finally executed.

See issue #3237.