Revert "libs: Update bundled QScintilla to version 2.13.3"
This reverts commit 6ddd4f019f2d3b4d2aaadc0cf9d2b62adbfd241f since it
breaks AppImage builds. The newer version of QScintilla depends on Qt
5.11 which is not available in the Ubuntu 18.04 the AppImages use. We
also cannot just use the version of QScintilla available in the OS
instead of the bundled one for the AppImages as that is too old for our
purposes. It is probably best to just update to Ubuntu 20.04 for the
AppImage builds in April 2023 when 18.04 reaches its end of life.