Updated the macOS nightly build scripts
Removed the NIGHTLY environment variable, because I realised when
editing these scripts that it would *always* be turned on.

We don't really have a use for the script where it's not a nightly
build. ;)

---

The builds themselves seem to go alright, except the updated
"nightly" icon file seems to make macOS unhappy.  That like needs
a bit more work done on it.