Workaround for issue #2652: release supposedly unneeded savepoints
By releasing the supposedly unneeded savepoints, we err on the safe side,
because if there were no modifications, the result is equivalent, and if
there were side effects on the select execution, we will save the changes
immediately, which, at least, will make these functions usable.

If there are schema changes, the user will have to refresh the DB manually,
though.