build: Move build artifacts from each folder
- github_actions/actions/download-artifacts-4
- github_actions/actions/upload-artifacts-4

Recently, GitHub updated these two actions to have different artifact
names for all build outputs, so if we set it to dnwload all artifacts on
download, it creates a folder for each artifact name and saves the build
outputs. so we change it to move the files saved in each folder to the
parent directory.

Related:
- 9fb0528faee5f75b4d7301878bb1977535fa0c97
- 447a2c779675ddb4235218e1ce44f68c718d38bf
- https://github.blog/changelog/2023-12-14-github-actions-artifacts-v4-is-now-generally-available/