Execute SQL: avoid problems with selection when there are multibyte chars
QScintilla returns selection in bytes, so we have to take into account
cases where the multibyte characters are inside the query, and also when
they are before the query. We do that by trimming after the end byte
position of the selection and replacing all the bytes previous to the
selection with a space (single-byte character)


See issue #3731.