Remove two key combinations using Ctrl+Alt+Letter
These are problematic on Windows, because they are translated to
AltGr+Letter and, consequently, can prevent insertion of a symbol,
depending on the keyboard layout.

Ctrl+Alt+C is still there, as I couldn't find an alternative using some
mnemonic, either with C or with Insert, which worked on Linux. Hopefully,
it won't affect users as it is only valid for the Data Browser (Copy as
SQL).

See issue #3771