Do not trust canConvert() because it only checks types, not values
This fixes the incorrect conversion of strings to 0 when the column is
of numeric affinity, like DATETIME with a string containing the
timestamp.

See comment in QTBUG-75776.
https://bugreports.qt.io/browse/QTBUG-75776?focusedId=459927&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-459927

> canConvert() only tests whether the type of the data held by the
QVariant is convertible to the specified type; it is only when convert()
attempts to do so that the actual data (of that type) gets looked at to
see whether it converts. This is the documented behaviour and the only
sensible way to design the API.

See issue #3785