Remove dbhub.io code
Remove all code related to dbhub.io because the servers have been shut
down some months ago. We can always restore this code later if needed.

This keeps some of the networking code as well as the proxy settings
because they are still needed for checking for new program versions.